Moved some stubs into a new file
